+++ src/Core/Type/Common/ParentType.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Gskema\TypeSniff\Core\Type\Common;
+
+use Gskema\TypeSniff\Core\Type\TypeInterface;
+
+class ParentType implements TypeInterface
+{
+    /**
+     * @inheritDoc
+     */
+    public function toString(): string
+    {
+        return 'parent';
+    }
+}